add order

+++ src/main/java/com/project/shopapp/dtos/OrderDetailDTO.java
@@ -13,14 +13,14 @@ import lombok.*;
 public class OrderDetailDTO {
     @JsonProperty("order_id")
     @Min(value = 1, message = "Order's ID must be > 0")
-    private int orderId;
+    private Long orderId;
 
     @JsonProperty("product_id")
     @Min(value = 1, message = "Product's ID must be > 0")
     private int productId;
 
     @Min(value = 0, message = "Product's Id must be >= 0")
-    private Long price;
+    private Float price;
 
     @JsonProperty("number_of_products")
     @Min(value = 1, message = "Number_of_products must be >= 1")
@@ -28,7 +28,7 @@ public class OrderDetailDTO {
 
     @JsonProperty("total_money")
     @Min(value = 0, message = "total_money must be >= 0")
-    private int totalMoney;
+    private Float totalMoney;
 
     private String color;
 }
